Modernized and updated all the embedded web authentication code. 1. Moved the embedded webview from UIWebView to WKWebView 2. Created a new authentication method (preferred) that a) fixes all Google auth related issues on both ios 9/10 and ios 11 b) still uses the embedded method for all other sign in types 3. Moved minimum supported version to ios 9 so autolayout can be used (also who uses 8?)
As for the google specific details, if on pre-ios 11, the new endpoint kicks the user out to safari. Google has denied embedded webview use for google sign in since april of 2017 and so the old endpoint has been failing for Google sign-in since then. for ios 11, the new endpoint uses new Apple API, namely `SFAuthenticationSession` for Google sign-in _only_.

One last note is that Apple has been rejecting my app for a) not having functioning Google sign-in and then b) for kicking the user out to Safari instead of an in-app sign in experience. Hopefully this is the best of both worlds.